+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Updated dependencies
     - scraper to 0.22.0
 - Added thiserror as dependency to refactor Error types
+- "Responsorial" will now be read as Psalm
+- Added "db store" command
 
 ## [0.3.2] - 2024-12-09
 - Code lint